Refactor Move all test classes in mocks directoy
