+++ src/mocks/providers/mock-camera.ts
